refactor(focus): extract FocusCandidateResolver from FocusSnapshotResolver
Move candidate search (resolveCandidate, candidateElements, appendEditableDescendants, looksEditable) and per-element AX reads (candidateSnapshot, nativeTextWindow, isSecureElement), plus the AXFocusCandidate / FocusCandidateResolution / AXTextSelection types, into a new FocusCandidateResolver. FocusSnapshotResolver is now a 268-line orchestrator (was 1020) that composes FocusCandidateResolver, CaretGeometryResolver, DeepGeometryWalkThrottle, and AXTreeDumpWriter.

Verbatim move; build, lint, and the full suite (777 tests) are green. Same manual AX smoke caveat as the rest of the focus split (native field, Chromium, Electron, secure field).
